EREvent.toJson returns a Ref, Mem.toJson returns an Option
As EREvent.toJson now returns a Ref, it can for instance have no JSON representation
or involve a Future (looking something up in the database)

Mem.toJson returns an option. The reason for not returning a Ref is that we really
don't want to be looking them up in the database, as member lists are sent out on
every join and quit request (which for a large eventroom could be very often).